requirements.txt
parsedatetime==2.6
Pillow==10.2.0
random_user_agent==1.0.1
Requests==2.31.0
selenium==4.18.1
tabulate==0.9.0
tqdm==4.66.1
webdriver_manager==4.0.0
